% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carre.R
\name{carre}
\alias{carre}
\title{Moyenne quadratique}
\usage{
carre(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Moyenne quadratique
}
\examples{
carre(3, 4)

}
